+++ dhis-2/dhis-support/dhis-support-json-filtering/src/test/java/org/hisp/dhis/commons/jsonfiltering/filter/SimpleJsonFilteringContextProvider.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2004-2021, University of Oslo
+ * Copyright (c) 2004-2004-2020, University of Oslo
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package squiggly.filter;
+package org.hisp.dhis.commons.jsonfiltering.filter;
 
 import org.hisp.dhis.commons.jsonfiltering.context.provider.AbstractJsonFilteringContextProvider;
 import org.hisp.dhis.commons.jsonfiltering.name.AnyDeepName;
