+++ dhis-2/dhis-web/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/export/PagingWrapperFieldFilterCustomizer.java
@@ -27,15 +27,15 @@
  */
 package org.hisp.dhis.webapi.controller.tracker.export;
 
+import java.util.Collection;
+import java.util.Collections;
+import java.util.regex.Pattern;
+
 import lombok.Getter;
 
 import org.hisp.dhis.commons.jsonfiltering.web.FieldFilterCustomizer;
 import org.hisp.dhis.webapi.controller.event.webrequest.PagingWrapper;
 import org.springframework.stereotype.Component;
-
-import java.util.Collection;
-import java.util.Collections;
-import java.util.regex.Pattern;
 
 @Component
 @SuppressWarnings( "rawtypes" )
@@ -43,8 +43,10 @@ class PagingWrapperFieldFilterCustomizer implements FieldFilterCustomizer<Paging
 {
     @Getter
     private final Class<PagingWrapper> applicableClass = PagingWrapper.class;
+
     @Getter
-    private final Collection<Pattern> supportedUriPatterns = Collections.singleton(Pattern.compile(".*/api/tracker.*"));
+    private final Collection<Pattern> supportedUriPatterns = Collections
+        .singleton( Pattern.compile( ".*/api/tracker.*" ) );
 
     @Override
     public String customize( String filter )
